Drop legacy log parsing
