+++ logs/legacy_parse.go
@@ -9,7 +9,6 @@ import (
 	"time"
 
 	"github.com/google/uuid"
-	"github.com/pganalyze/collector/output/pganalyze_collector"
 	"github.com/pganalyze/collector/state"
 )
 
@@ -103,399 +102,6 @@ var RsyslogTimeRegexp = `(\w+\s+\d+ \d{2}:\d{2}:\d{2})`
 var RsyslogHostnameRegxp = `(\S+)`
 var RsyslogProcessNameRegexp = `(\w+)`
 var RsyslogRegexp = regexp.MustCompile(`^` + RsyslogTimeRegexp + ` ` + RsyslogHostnameRegxp + ` ` + RsyslogProcessNameRegexp + `\[` + PidRegexp + `\]: ` + SyslogSequenceAndSplitRegexp + ` ` + RsyslogLevelAndContentRegexp)
-
-func ParseLogLineWithPrefix(prefix string, line string, tz *time.Location) (logLine state.LogLine, ok bool) {
-	var timePart, userPart, dbPart, appPart, pidPart, logLineNumberPart, levelPart, contentPart string
-
-	rsyslog := false
-
-	// Only read the first 1000 characters of a log line to parse the log_line_prefix
-	//
-	// This reduces the overhead for very long loglines, because we don't pass in the
-	// whole line to the regexp engine (twice).
-	lineExtra := ""
-	if len(line) > 1000 {
-		lineExtra = line[1000:]
-		line = line[0:1000]
-	}
-
-	if prefix == "" {
-		if LogPrefixAmazonRdsRegexp.MatchString(line) {
-			prefix = LogPrefixAmazonRds
-		} else if LogPrefixAzureRegexp.MatchString(line) {
-			prefix = LogPrefixAzure
-		} else if LogPrefixCustom1Regexp.MatchString(line) {
-			prefix = LogPrefixCustom1
-		} else if LogPrefixCustom2Regexp.MatchString(line) {
-			prefix = LogPrefixCustom2
-		} else if LogPrefixCustom4Regexp.MatchString(line) { // 4 is more specific than 3, so needs to go first
-			prefix = LogPrefixCustom4
-		} else if LogPrefixCustom3Regexp.MatchString(line) {
-			prefix = LogPrefixCustom3
-		} else if LogPrefixCustom5Regexp.MatchString(line) {
-			prefix = LogPrefixCustom5
-		} else if LogPrefixCustom6Regexp.MatchString(line) {
-			prefix = LogPrefixCustom6
-		} else if LogPrefixCustom7Regexp.MatchString(line) {
-			prefix = LogPrefixCustom7
-		} else if LogPrefixCustom8Regexp.MatchString(line) {
-			prefix = LogPrefixCustom8
-		} else if LogPrefixCustom9Regexp.MatchString(line) {
-			prefix = LogPrefixCustom9
-		} else if LogPrefixCustom10Regexp.MatchString(line) {
-			prefix = LogPrefixCustom10
-		} else if LogPrefixCustom11Regexp.MatchString(line) {
-			prefix = LogPrefixCustom11
-		} else if LogPrefixCustom12Regexp.MatchString(line) {
-			prefix = LogPrefixCustom12
-		} else if LogPrefixCustom13Regexp.MatchString(line) {
-			prefix = LogPrefixCustom13
-		} else if LogPrefixCustom14Regexp.MatchString(line) {
-			prefix = LogPrefixCustom14
-		} else if LogPrefixCustom15Regexp.MatchString(line) {
-			prefix = LogPrefixCustom15
-		} else if LogPrefixCustom16Regexp.MatchString(line) {
-			prefix = LogPrefixCustom16
-		} else if LogPrefixSimpleRegexp.MatchString(line) {
-			prefix = LogPrefixSimple
-		} else if LogPrefixHeroku2Regexp.MatchString(line) {
-			prefix = LogPrefixHeroku2
-		} else if LogPrefixHeroku1Regexp.MatchString(line) {
-			// LogPrefixHeroku1 is a subset of 2, so it must be matched second
-			prefix = LogPrefixHeroku1
-		} else if RsyslogRegexp.MatchString(line) {
-			rsyslog = true
-		}
-	}
-
-	if rsyslog {
-		parts := RsyslogRegexp.FindStringSubmatch(line)
-		if len(parts) == 0 {
-			return
-		}
-
-		timePart = fmt.Sprintf("%d %s", time.Now().Year(), parts[1])
-		// ignore syslog hostname
-		// ignore syslog process name
-		pidPart = parts[4]
-		// ignore syslog postgres sequence and split number
-		levelPart = parts[6]
-		contentPart = strings.Replace(parts[7], "#011", "\t", -1)
-
-		parts = LogPrefixNoTimestampUserDatabaseAppRegexp.FindStringSubmatch(contentPart)
-		if len(parts) == 6 {
-			userPart = parts[1]
-			dbPart = parts[2]
-			appPart = parts[3]
-			levelPart = parts[4]
-			contentPart = parts[5]
-		}
-	} else {
-		switch prefix {
-		case LogPrefixAmazonRds: // "%t:%r:%u@%d:[%p]:"
-			parts := LogPrefixAmazonRdsRegexp.FindStringSubmatch(line)
-			if len(parts) == 0 {
-				return
-			}
-
-			timePart = parts[1]
-			// skip %r (ip+port)
-			userPart = parts[3]
-			dbPart = parts[4]
-			pidPart = parts[5]
-			levelPart = parts[6]
-			contentPart = parts[7]
-		case LogPrefixAzure: // "%t-%c-"
-			parts := LogPrefixAzureRegexp.FindStringSubmatch(line)
-			if len(parts) == 0 {
-				return
-			}
-
-			timePart = parts[1]
-			// skip %c (session id)
-			levelPart = parts[3]
-			contentPart = parts[4]
-		case LogPrefixCustom1: // "%m [%p][%v] : [%l-1] %q[app=%a] "
-			parts := LogPrefixCustom1Regexp.FindStringSubmatch(line)
-			if len(parts) == 0 {
-				return
-			}
-			timePart = parts[1]
-			pidPart = parts[2]
-			// skip %v (virtual TX)
-			logLineNumberPart = parts[4]
-			appPart = parts[5]
-			levelPart = parts[6]
-			contentPart = parts[7]
-		case LogPrefixCustom2: // "%t [%p-1] %q%u@%d "
-			parts := LogPrefixCustom2Regexp.FindStringSubmatch(line)
-			if len(parts) == 0 {
-				return
-			}
-			timePart = parts[1]
-			pidPart = parts[2]
-			logLineNumberPart = parts[3]
-			userPart = parts[4]
-			dbPart = parts[5]
-			levelPart = parts[6]
-			contentPart = parts[7]
-		case LogPrefixCustom3: // "%m [%p] %q[user=%u,db=%d,app=%a] ""
-			parts := LogPrefixCustom3Regexp.FindStringSubmatch(line)
-			if len(parts) == 0 {
-				return
-			}
-			timePart = parts[1]
-			pidPart = parts[2]
-			userPart = parts[3]
-			dbPart = parts[4]
-			appPart = parts[5]
-			levelPart = parts[6]
-			contentPart = parts[7]
-		case LogPrefixCustom4: // "%m [%p] %q[user=%u,db=%d,app=%a,host=%h] "
-			parts := LogPrefixCustom4Regexp.FindStringSubmatch(line)
-			if len(parts) == 0 {
-				return
-			}
-			timePart = parts[1]
-			pidPart = parts[2]
-			userPart = parts[3]
-			dbPart = parts[4]
-			appPart = parts[5]
-			// skip %h (host)
-			levelPart = parts[7]
-			contentPart = parts[8]
-		case LogPrefixCustom5: // "%t [%p]: [%l-1] user=%u,db=%d - PG-%e "
-			parts := LogPrefixCustom5Regexp.FindStringSubmatch(line)
-			if len(parts) == 0 {
-				return
-			}
-			timePart = parts[1]
-			pidPart = parts[2]
-			logLineNumberPart = parts[3]
-			userPart = parts[4]
-			dbPart = parts[5]
-			// skip %e (SQLSTATE)
-			levelPart = parts[7]
-			contentPart = parts[8]
-		case LogPrefixCustom6: // "%t [%p]: [%l-1] user=%u,db=%d,app=%a,client=%h "
-			parts := LogPrefixCustom6Regexp.FindStringSubmatch(line)
-			if len(parts) == 0 {
-				return
-			}
-			timePart = parts[1]
-			pidPart = parts[2]
-			logLineNumberPart = parts[3]
-			userPart = parts[4]
-			dbPart = parts[5]
-			// skip %a (application name)
-			// skip %h (host)
-			levelPart = parts[8]
-			contentPart = parts[9]
-		case LogPrefixCustom7: // "%t [%p]: [%l-1] [trx_id=%x] user=%u,db=%d "
-			parts := LogPrefixCustom7Regexp.FindStringSubmatch(line)
-			if len(parts) == 0 {
-				return
-			}
-			timePart = parts[1]
-			pidPart = parts[2]
-			logLineNumberPart = parts[3]
-			// skip %x (transaction id)
-			userPart = parts[5]
-			dbPart = parts[6]
-			levelPart = parts[7]
-			contentPart = parts[8]
-		case LogPrefixCustom8: // "[%p]: [%l-1] db=%d,user=%u "
-			parts := LogPrefixCustom8Regexp.FindStringSubmatch(line)
-			if len(parts) == 0 {
-				return
-			}
-			pidPart = parts[1]
-			logLineNumberPart = parts[2]
-			dbPart = parts[3]
-			userPart = parts[4]
-			levelPart = parts[5]
-			contentPart = parts[6]
-		case LogPrefixCustom9: // "%m %r %u %a [%c] [%p] "
-			parts := LogPrefixCustom9Regexp.FindStringSubmatch(line)
-			if len(parts) == 0 {
-				return
-			}
-			timePart = parts[1]
-			// skip %r (ip+port)
-			userPart = parts[3]
-			appPart = parts[4]
-			// skip %c (session id)
-			pidPart = parts[6]
-			levelPart = parts[7]
-			contentPart = parts[8]
-		case LogPrefixCustom10: // "%t [%p]: [%l-1] db=%d,user=%u "
-			parts := LogPrefixCustom10Regexp.FindStringSubmatch(line)
-			if len(parts) == 0 {
-				return
-			}
-			timePart = parts[1]
-			pidPart = parts[2]
-			logLineNumberPart = parts[3]
-			dbPart = parts[4]
-			userPart = parts[5]
-			levelPart = parts[6]
-			contentPart = parts[7]
-		case LogPrefixCustom11: // "pid=%p,user=%u,db=%d,app=%a,client=%h "
-			parts := LogPrefixCustom11Regexp.FindStringSubmatch(line)
-			if len(parts) == 0 {
-				return
-			}
-			pidPart = parts[1]
-			userPart = parts[2]
-			dbPart = parts[3]
-			// skip %a (application name)
-			// skip %h (host)
-			levelPart = parts[6]
-			contentPart = parts[7]
-		case LogPrefixCustom12: // "user=%u,db=%d,app=%a,client=%h "
-			parts := LogPrefixCustom12Regexp.FindStringSubmatch(line)
-			if len(parts) == 0 {
-				return
-			}
-			userPart = parts[1]
-			dbPart = parts[2]
-			// skip %a (application name)
-			// skip %h (host)
-			levelPart = parts[5]
-			contentPart = parts[6]
-		case LogPrefixCustom13: // "%p-%s-%c-%l-%h-%u-%d-%m "
-			parts := LogPrefixCustom13Regexp.FindStringSubmatch(line)
-			if len(parts) == 0 {
-				return
-			}
-			pidPart = parts[1]
-			// skip %s
-			// skip %c
-			logLineNumberPart = parts[4]
-			// skip %h (host)
-			userPart = parts[6]
-			dbPart = parts[7]
-			timePart = parts[8]
-			levelPart = parts[9]
-			contentPart = parts[10]
-		case LogPrefixCustom14: // "%m [%p][%b][%v][%x] %q[user=%u,db=%d,app=%a] "
-			parts := LogPrefixCustom14Regexp.FindStringSubmatch(line)
-			if len(parts) == 0 {
-				return
-			}
-
-			timePart = parts[1]
-			pidPart = parts[2]
-			// skip %b
-			// skip %v
-			// skip %x
-			userPart = parts[6]
-			dbPart = parts[7]
-			appPart = parts[8]
-			levelPart = parts[9]
-			contentPart = parts[10]
-		case LogPrefixCustom15: // "%m [%p] %q%u@%d "
-			parts := LogPrefixCustom15Regexp.FindStringSubmatch(line)
-			if len(parts) == 0 {
-				return
-			}
-			timePart = parts[1]
-			pidPart = parts[2]
-			userPart = parts[3]
-			dbPart = parts[4]
-			levelPart = parts[5]
-			contentPart = parts[6]
-		case LogPrefixCustom16: // "%t [%p] %q%u@%d %h "
-			parts := LogPrefixCustom16Regexp.FindStringSubmatch(line)
-			if len(parts) == 0 {
-				return
-			}
-			timePart = parts[1]
-			pidPart = parts[2]
-			userPart = parts[3]
-			dbPart = parts[4]
-			// skip %h (host)
-			levelPart = parts[6]
-			contentPart = parts[7]
-		case LogPrefixSimple: // "%t [%p] "
-			parts := LogPrefixSimpleRegexp.FindStringSubmatch(line)
-			if len(parts) == 0 {
-				return
-			}
-			timePart = parts[1]
-			pidPart = parts[2]
-			levelPart = parts[3]
-			contentPart = parts[4]
-		case LogPrefixHeroku1:
-			parts := LogPrefixHeroku1Regexp.FindStringSubmatch(line)
-			if len(parts) == 0 {
-				return
-			}
-			// skip %e
-			levelPart = parts[2]
-			contentPart = parts[3]
-		case LogPrefixHeroku2:
-			parts := LogPrefixHeroku2Regexp.FindStringSubmatch(line)
-			if len(parts) == 0 {
-				return
-			}
-			// skip %e
-			timePart = parts[2]
-			pidPart = parts[3]
-			// skip %s
-			// skip %c
-			// skip %v
-			// skip %x
-			logLineNumberPart = parts[8]
-			dbPart = parts[9]
-			// skip %r
-			userPart = parts[11]
-			appPart = parts[12]
-			levelPart = parts[13]
-			contentPart = parts[14]
-		default:
-			// Some callers use the content of unparsed lines to stitch multi-line logs together
-			logLine.Content = line + lineExtra
-			return
-		}
-	}
-
-	if timePart != "" {
-		occurredAt := GetOccurredAt(timePart, tz, rsyslog)
-		if occurredAt.IsZero() {
-			return
-		}
-		logLine.OccurredAt = occurredAt
-	}
-
-	if userPart != "[unknown]" {
-		logLine.Username = userPart
-	}
-	if dbPart != "[unknown]" {
-		logLine.Database = dbPart
-	}
-	if appPart != "[unknown]" {
-		logLine.Application = appPart
-	}
-	if logLineNumberPart != "" {
-		logLineNumber, _ := strconv.ParseInt(logLineNumberPart, 10, 32)
-		logLine.LogLineNumber = int32(logLineNumber)
-	}
-
-	backendPid, _ := strconv.ParseInt(pidPart, 10, 32)
-	logLine.BackendPid = int32(backendPid)
-	logLine.Content = contentPart + lineExtra
-
-	// This is actually a continuation of a previous line
-	if levelPart == "" {
-		return
-	}
-
-	logLine.LogLevel = pganalyze_collector.LogLineInformation_LogLevel(pganalyze_collector.LogLineInformation_LogLevel_value[levelPart])
-	ok = true
-
-	return
-}
 
 func GetOccurredAt(timePart string, tz *time.Location, rsyslog bool) time.Time {
 	if tz != nil && !rsyslog {
